Move codegen code into parser_utils
